Remove extracting progress to external file
This would make it more difficult to reason about `SEXP bar` in particular, as we'd not know it is protected already, and that's a very bad pattern.
